resources/requirements.txt
discord.py==1.7.3
feedparser==6.0.8
python-dotenv==0.19.2
